feat: add page connect #23
