+++ app/src/main/java/com/example/mnrader/addScreens/SelectTypeScreen.kt
@@ -19,13 +19,19 @@ import androidx.compose.ui.tooling.preview.Preview
 import androidx.compose.ui.unit.dp
 import androidx.compose.ui.unit.sp
 import androidx.navigation.NavController
+import androidx.navigation.NavHostController
 import androidx.navigation.compose.rememberNavController
 import com.example.mnrader.model.RegisterScreens
 import com.example.mnrader.model.RegisterViewModel
 import com.example.mnrader.navigation.RegisterTopBar
+import com.example.mnrader.navigation.Routes
 
 @Composable
-fun SelectTypeScreen(navController: NavController, viewModel: RegisterViewModel) {
+fun SelectTypeScreen(
+    navController: NavController,
+    rootNavController: NavHostController,
+    viewModel: RegisterViewModel
+) {
 
     val customButtonColor = Color(0xFF89C5A9)
 
@@ -33,10 +39,8 @@ fun SelectTypeScreen(navController: NavController, viewModel: RegisterViewModel)
         modifier = Modifier.fillMaxSize()) {
         RegisterTopBar(
             onBackClick = {//Home화면으로 이동
-//                navController.popBackStack()
-                navController.navigate("HomeScreen") {
-                    popUpTo("RegisterStartScreen") { inclusive = true }
-                } },
+                rootNavController.popBackStack()  // 또는 rootNavController.navigate(Routes.MAIN)
+                },
             currentStep = 1 // 여기서 단계 조정: 1~5
         )
         Box(modifier = Modifier.fillMaxSize()
@@ -79,5 +83,5 @@ fun SelectTypeScreenPreview() {
     val navController = rememberNavController()
     val viewModel = remember { RegisterViewModel() }
 
-    SelectTypeScreen(navController = navController, viewModel = viewModel)
+    SelectTypeScreen(navController = navController, rootNavController = navController,viewModel = viewModel)
 }
