+++ app/src/main/java/com/example/mnrader/navigation/MainNavHost.kt
@@ -7,6 +7,14 @@ import androidx.compose.ui.Modifier
 import androidx.navigation.NavHostController
 import androidx.navigation.compose.NavHost
 import androidx.navigation.compose.composable
+import androidx.navigation.compose.rememberNavController
+import com.example.mnrader.addScreens.AnimalTypeScreen
+import com.example.mnrader.addScreens.RegisterInfoScreen
+import com.example.mnrader.addScreens.ReportOrLostScreen
+import com.example.mnrader.addScreens.SelectTypeScreen
+import com.example.mnrader.addScreens.SubmitSuccessScreen
+import com.example.mnrader.model.RegisterScreens
+import com.example.mnrader.model.RegisterViewModel
 import com.example.mnrader.ui.home.screen.HomeScreen
 import com.example.mnrader.ui.onboarding.screen.OnboardingScreen
 
@@ -72,7 +80,9 @@ fun MainNavHost(
         // 등록하기
         composable(
             route = Routes.ADD
-        ) { }
+        ) {
+            AnimalRegister(navController)
+        }
 
         // 마이페이지
         composable(
@@ -97,4 +107,34 @@ object Routes {
     const val ADD = "add"
     const val MYPAGE = "mypage"
     const val SETTING = "setting"
+}
+
+
+@Composable
+fun AnimalRegister(
+    rootNavController: NavHostController,
+    viewModel: RegisterViewModel = RegisterViewModel()
+) {
+    val navController = rememberNavController()
+
+    NavHost(
+        navController = navController,
+        startDestination = RegisterScreens.SelectType.route
+    ) {
+        composable(RegisterScreens.SelectType.route) {
+            SelectTypeScreen(navController,rootNavController, viewModel)
+        }
+        composable(RegisterScreens.RegisterInfo.route) {
+            RegisterInfoScreen(navController, viewModel)
+        }
+        composable(RegisterScreens.AnimalType.route) {
+            AnimalTypeScreen(navController, viewModel)
+        }
+        composable(RegisterScreens.ReportOrLost.route) {
+            ReportOrLostScreen(navController, viewModel)
+        }
+        composable(RegisterScreens.SubmitSuccess.route) {
+            SubmitSuccessScreen(navController, rootNavController,viewModel)
+        }
+    }
 }
